add checks & validation

+++ app/Http/Controllers/UsersController.php
@@ -10,19 +10,6 @@ class UsersController extends Controller
 {
 
 
-    /**
-     * Create a new controller instance.
-     *
-     * @return void
-     */
-    public function __construct()
-    {
-        //
-    }
-
-    
-
-
 
     public function index()
     {
@@ -38,6 +25,9 @@ class UsersController extends Controller
 
     public function show($userId)
     {
+        if(!$this->checkValidIdlength($userId)) 
+            return $this->invalidIdResponse(); 
+
         return response()->json([
             "data" => User::findById($userId), 
         ], 200); 
@@ -48,6 +38,16 @@ class UsersController extends Controller
 
     public function store(Request $request, $userId = null)
     {
+        $this->validate($request, [
+            'username' => 'required',
+            'email' => 'required',
+            'role' => 'required'
+        ]); 
+
+        if(!is_null($userId) && !$this->checkValidIdlength($userId))
+            return $this->invalidIdResponse(); 
+
+
         return response()->json([
             "data" => User::persist(
                 $request->all(["username","email","role"]), 
@@ -59,27 +59,35 @@ class UsersController extends Controller
 
 
 
-    public function update(Request $request, $userId)
-    {
-        return response()->json([
-            "data" => User::findAndUpdate(
-                $userId, 
-                $request->all(["username","email","role"])
-            )
-        ]); 
-    }
-
-
-
 
     public function destroy(Request $request, $userId)
     {
+        if(!$this->checkValidIdlength($userId)) 
+            return $this->invalidIdResponse(); 
 
         return response()->json([
             "data" => User::delete($userId)
         ]); 
 
 
+    }
+
+
+
+
+    protected function checkValidIdlength($id)
+    {
+        return strlen($id) == 24; 
+    }
+
+
+
+
+    protected function invalidIdResponse()
+    {
+        return response()->json([
+            "error" => "Incorrect Id"
+        ], 400); 
     }
